♻️ [Refactoring]: split plugin/ cache implementations into plugin/cache/ subfolder
Move cache.rs / cached_package.rs / cleanup.rs into plugin/cache/ and aggregate
their public types via plugin/cache.rs (sub-parent). Re-exports follow upstream
visibility: pub(crate) use for list_installed and cleanup_* (originally
pub(crate) fn), pub use for PackageCache, PackageCacheAccess, CachedPackage.
Apply #[allow(clippy::module_inception)] to the same-name leaf module. Rewrite
the cross-subgroup super::* references in cache.rs to crate::plugin::* paths
(including super::meta::write_installed_at -> crate::plugin::meta::write_installed_at,
ref 論点 e). Rewrite marketplace_content.rs:13 and marketplace_content_test.rs:4
since the cached_package module path is removed in this commit.

Refs #246
