+++ src/plugin.rs
@@ -1,7 +1,5 @@
 mod action;
 mod cache;
-mod cached_package;
-mod cleanup;
 mod installed;
 mod intent;
 mod loader;
@@ -14,9 +12,8 @@ mod update;
 mod version;
 
 pub use action::PluginAction;
-pub(crate) use cache::list_installed;
+pub(crate) use cache::{cleanup_legacy_hierarchy, cleanup_plugin_directories, list_installed};
 pub use cache::{CachedPackage, PackageCache, PackageCacheAccess};
-pub(crate) use cleanup::{cleanup_legacy_hierarchy, cleanup_plugin_directories};
 pub use installed::InstalledPlugin;
 pub use intent::PluginIntent;
 pub(crate) use loader::load_plugin;
