+++ src/plugin/cache/cache.rs
@@ -1,0 +1,682 @@
+//! パッケージキャッシュマネージャ
+//!
+//! GitHubやマーケットプレイスからダウンロードしたパッケージのキャッシュ管理を行う。
+
+use crate::error::{PlmError, Result};
+use crate::fs::{FileSystem, RealFs};
+use crate::plugin::manifest_resolve::resolve_manifest_path;
+use crate::plugin::MarketplaceContent;
+use crate::plugin::{meta, PluginManifest};
+use std::io::{Cursor, Read};
+use std::path::{Component as PathComponent, Path, PathBuf};
+use zip::ZipArchive;
+
+// Re-export
+pub use super::cached_package::CachedPackage;
+pub use crate::plugin::manifest_resolve::has_manifest;
+
+/// cache.list() の生タプルからのマッピング型（変更吸収層）
+struct PluginCacheKey {
+    marketplace: Option<String>,
+    name: String,
+}
+
+impl From<(Option<String>, String)> for PluginCacheKey {
+    fn from((marketplace, name): (Option<String>, String)) -> Self {
+        Self { marketplace, name }
+    }
+}
+
+/// キャッシュ内のマーケットプレイスパッケージを列挙
+///
+/// # Arguments
+///
+/// * `cache` - package cache access used to enumerate stored plugins
+pub(crate) fn list_installed(cache: &dyn PackageCacheAccess) -> Result<Vec<MarketplaceContent>> {
+    // 一覧取得経路では個別プラグインのスキャン失敗を握りつぶし、列挙自体は続行する。
+    // TUI 経由でも呼ばれるため stderr への直接出力は避ける。スキャン失敗を厳密に
+    // 検出する必要のある経路（install 等）は `Plugin::new` / `try_from` を直接呼ぶこと。
+    let packages = cache
+        .list()?
+        .into_iter()
+        .map(PluginCacheKey::from)
+        .filter(|key| !key.name.starts_with('.'))
+        .filter_map(|key| {
+            cache
+                .load_package(key.marketplace.as_deref(), &key.name)
+                .ok()
+                .and_then(|pkg| MarketplaceContent::try_from(pkg).ok())
+        })
+        .collect();
+    Ok(packages)
+}
+
+/// パッケージキャッシュアクセスの抽象化トレイト
+///
+/// 消費者はこの trait 経由でキャッシュ操作を行う。
+/// テスト時は `PackageCache::with_cache_dir(tempdir)` で tempdir ベースの
+/// 本番実装を注入する。
+pub trait PackageCacheAccess: Send + Sync {
+    /// プラグインのキャッシュパスを取得（階層型: marketplace/plugin）
+    ///
+    /// # Arguments
+    ///
+    /// * `marketplace` - マーケットプレイス名（None の場合は "github" を使用）
+    /// * `name` - プラグイン名またはリポジトリ識別子（owner--repo 形式）
+    fn plugin_path(&self, marketplace: Option<&str>, name: &str) -> PathBuf;
+
+    /// キャッシュ済みかチェック
+    ///
+    /// # Arguments
+    ///
+    /// * `marketplace` - marketplace name (`None` uses `"github"`)
+    /// * `name` - plugin name or repository identifier
+    fn is_cached(&self, marketplace: Option<&str>, name: &str) -> bool;
+
+    /// zipアーカイブを展開してキャッシュに保存
+    ///
+    /// GitHubのzipballは `{repo}-{ref}/` というプレフィックスが付くため、それを除去する。
+    ///
+    /// # Arguments
+    ///
+    /// * `marketplace` - マーケットプレイス名（None の場合は "github" を使用）
+    /// * `name` - プラグイン名またはリポジトリ識別子
+    /// * `archive` - zipアーカイブのバイト列
+    /// * `source_path` - 抽出するソースパス（正規化済み、例: "plugins/my-plugin"）
+    ///   指定時はそのパス配下の内容のみをキャッシュ直下に展開
+    fn store_from_archive(
+        &self,
+        marketplace: Option<&str>,
+        name: &str,
+        archive: &[u8],
+        source_path: Option<&str>,
+    ) -> Result<PathBuf>;
+
+    /// キャッシュからマニフェストを読み込み
+    ///
+    /// # Arguments
+    ///
+    /// * `marketplace` - marketplace name (`None` uses `"github"`)
+    /// * `name` - plugin name or repository identifier
+    fn load_manifest(&self, marketplace: Option<&str>, name: &str) -> Result<PluginManifest>;
+
+    /// キャッシュから削除
+    ///
+    /// # Arguments
+    ///
+    /// * `marketplace` - marketplace name (`None` uses `"github"`)
+    /// * `name` - plugin name or repository identifier
+    fn remove(&self, marketplace: Option<&str>, name: &str) -> Result<()>;
+
+    /// キャッシュされているプラグイン一覧を取得
+    fn list(&self) -> Result<Vec<(Option<String>, String)>>;
+
+    /// プラグインをバックアップ
+    ///
+    /// 注意: 実行ビット/シンボリックリンクは保持されない。
+    /// プラグインは平文ファイルのみを含むことを前提とする。
+    ///
+    /// # Arguments
+    ///
+    /// * `marketplace` - marketplace name (`None` uses `"github"`)
+    /// * `name` - plugin name or repository identifier
+    fn backup(&self, marketplace: Option<&str>, name: &str) -> Result<PathBuf>;
+
+    /// バックアップからリストア
+    ///
+    /// # Arguments
+    ///
+    /// * `marketplace` - marketplace name (`None` uses `"github"`)
+    /// * `name` - plugin name or repository identifier
+    fn restore(&self, marketplace: Option<&str>, name: &str) -> Result<()>;
+
+    /// バックアップを削除
+    ///
+    /// # Arguments
+    ///
+    /// * `marketplace` - marketplace name (`None` uses `"github"`)
+    /// * `name` - plugin name or repository identifier
+    fn remove_backup(&self, marketplace: Option<&str>, name: &str) -> Result<()>;
+
+    /// キャッシュからCachedPackageを生成
+    ///
+    /// plugin_path + load_manifest + load_meta を組み合わせて
+    /// CachedPackage を構築する。
+    ///
+    /// # Arguments
+    ///
+    /// * `marketplace` - marketplace name (`None` uses `"github"`)
+    /// * `name` - plugin name or repository identifier
+    fn load_package(&self, marketplace: Option<&str>, name: &str) -> Result<CachedPackage> {
+        use super::cached_package::UNKNOWN_GIT_VALUE;
+
+        let path = self.plugin_path(marketplace, name);
+        let manifest = self.load_manifest(marketplace, name)?;
+        let loaded_meta = meta::load_meta(&path);
+        let git_ref = loaded_meta
+            .as_ref()
+            .and_then(|m| m.git_ref.clone())
+            .unwrap_or_else(|| UNKNOWN_GIT_VALUE.to_string());
+        let commit_sha = loaded_meta
+            .as_ref()
+            .and_then(|m| m.commit_sha.clone())
+            .unwrap_or_else(|| UNKNOWN_GIT_VALUE.to_string());
+
+        Ok(CachedPackage {
+            name: manifest.name.clone(),
+            id: Some(name.to_string()),
+            marketplace: marketplace.map(String::from),
+            path,
+            manifest,
+            git_ref,
+            commit_sha,
+            marketplace_manifest: None,
+        })
+    }
+
+    /// アトミック更新（temp展開 → 検証 → リネーム）
+    ///
+    /// 中断耐性のあるキャッシュ更新を行う。
+    /// 展開ロジックは store_from_archive と同一（トップディレクトリ除去含む）。
+    ///
+    /// # Arguments
+    ///
+    /// * `marketplace` - マーケットプレイス名
+    /// * `name` - プラグイン名
+    /// * `archive` - zipアーカイブのバイト列
+    fn atomic_update(
+        &self,
+        marketplace: Option<&str>,
+        name: &str,
+        archive: &[u8],
+    ) -> Result<PathBuf>;
+}
+
+/// パッケージキャッシュマネージャ
+pub struct PackageCache {
+    /// キャッシュルート: `$PLM_HOME/.plm/cache/plugins/` または `~/.plm/cache/plugins/`
+    cache_dir: PathBuf,
+}
+
+impl PackageCache {
+    /// キャッシュマネージャを初期化（ディレクトリ作成含む）
+    ///
+    /// `PLM_HOME` が設定されている場合はそちらを優先し、なければ `HOME` にフォールバックする。
+    pub fn new() -> Result<Self> {
+        let fs = RealFs;
+        let home = crate::env::EnvVar::get("PLM_HOME")
+            .or_else(|| crate::env::EnvVar::get("HOME"))
+            .ok_or_else(|| {
+                PlmError::Cache(
+                    "PLM_HOME and HOME environment variables not set or empty".to_string(),
+                )
+            })?;
+        let cache_dir = PathBuf::from(home)
+            .join(".plm")
+            .join("cache")
+            .join("plugins");
+        fs.create_dir_all(&cache_dir)?;
+        Ok(Self { cache_dir })
+    }
+
+    /// カスタムキャッシュディレクトリで初期化（テスト用）
+    ///
+    /// # Arguments
+    ///
+    /// * `cache_dir` - custom cache root directory
+    pub fn with_cache_dir(cache_dir: PathBuf) -> Result<Self> {
+        let fs = RealFs;
+        fs.create_dir_all(&cache_dir)?;
+        Ok(Self { cache_dir })
+    }
+
+    /// 全キャッシュをクリア
+    pub fn clear(&self) -> Result<()> {
+        let fs = RealFs;
+        if fs.exists(&self.cache_dir) {
+            for entry in fs.read_dir(&self.cache_dir)? {
+                if entry.is_dir() {
+                    fs.remove_dir_all(&entry.path)?;
+                }
+            }
+        }
+        Ok(())
+    }
+
+    /// バックアップパスを取得
+    ///
+    /// # Arguments
+    ///
+    /// * `marketplace` - marketplace name (`None` uses `"github"`)
+    /// * `name` - plugin name or repository identifier
+    fn backup_path(&self, marketplace: Option<&str>, name: &str) -> PathBuf {
+        let marketplace_dir = marketplace.unwrap_or("github");
+        self.cache_dir
+            .join(".backup")
+            .join(marketplace_dir)
+            .join(name)
+    }
+
+    /// 一時パスを取得
+    ///
+    /// # Arguments
+    ///
+    /// * `marketplace` - marketplace name (`None` uses `"github"`)
+    /// * `name` - plugin name or repository identifier
+    fn temp_path(&self, marketplace: Option<&str>, name: &str) -> PathBuf {
+        let marketplace_dir = marketplace.unwrap_or("github");
+        self.cache_dir
+            .join(".temp")
+            .join(marketplace_dir)
+            .join(name)
+    }
+}
+
+impl PackageCacheAccess for PackageCache {
+    fn plugin_path(&self, marketplace: Option<&str>, name: &str) -> PathBuf {
+        let marketplace_dir = marketplace.unwrap_or("github");
+        self.cache_dir.join(marketplace_dir).join(name)
+    }
+
+    fn is_cached(&self, marketplace: Option<&str>, name: &str) -> bool {
+        let fs = RealFs;
+        let plugin_path = self.plugin_path(marketplace, name);
+        fs.exists(&plugin_path)
+    }
+
+    fn store_from_archive(
+        &self,
+        marketplace: Option<&str>,
+        name: &str,
+        archive: &[u8],
+        source_path: Option<&str>,
+    ) -> Result<PathBuf> {
+        let fs = RealFs;
+
+        // source_path の防御的検証
+        validate_source_path(source_path)?;
+
+        let plugin_dir = self.plugin_path(marketplace, name);
+
+        // 既存のキャッシュがあれば削除
+        if fs.exists(&plugin_dir) {
+            fs.remove_dir_all(&plugin_dir)?;
+        }
+
+        // アーカイブを展開
+        extract_archive_with_source_path(&plugin_dir, archive, source_path)?;
+
+        // インストール日時を .plm-meta.json に記録（失敗時は警告のみ、処理は継続）
+        if let Err(e) = crate::plugin::meta::write_installed_at(&plugin_dir) {
+            eprintln!("Warning: Failed to write installedAt: {}", e);
+        }
+
+        Ok(plugin_dir)
+    }
+
+    fn load_manifest(&self, marketplace: Option<&str>, name: &str) -> Result<PluginManifest> {
+        let plugin_dir = self.plugin_path(marketplace, name);
+        let manifest_path = resolve_manifest_path(&plugin_dir).ok_or_else(|| {
+            PlmError::InvalidManifest(format!("plugin.json not found in {:?}", plugin_dir))
+        })?;
+
+        PluginManifest::load(&manifest_path)
+    }
+
+    fn remove(&self, marketplace: Option<&str>, name: &str) -> Result<()> {
+        let fs = RealFs;
+        let plugin_dir = self.plugin_path(marketplace, name);
+        if fs.exists(&plugin_dir) {
+            fs.remove_dir_all(&plugin_dir)?;
+        }
+        Ok(())
+    }
+
+    fn list(&self) -> Result<Vec<(Option<String>, String)>> {
+        let fs = RealFs;
+        let mut plugins = Vec::new();
+
+        if !fs.exists(&self.cache_dir) {
+            return Ok(plugins);
+        }
+
+        // marketplace ディレクトリを走査
+        for mp_entry in fs.read_dir(&self.cache_dir)? {
+            if !mp_entry.is_dir() {
+                continue;
+            }
+
+            let mp_path = &mp_entry.path;
+            let marketplace_name = mp_path
+                .file_name()
+                .and_then(|n| n.to_str())
+                .map(String::from);
+
+            // marketplace 内のプラグインを走査
+            for plugin_entry in fs.read_dir(mp_path)? {
+                if plugin_entry.is_dir() {
+                    if let Some(plugin_name) = plugin_entry.path.file_name() {
+                        let plugin_name = plugin_name.to_string_lossy().to_string();
+                        // "github" marketplace は None として扱う
+                        let mp = if marketplace_name.as_deref() == Some("github") {
+                            None
+                        } else {
+                            marketplace_name.clone()
+                        };
+                        plugins.push((mp, plugin_name));
+                    }
+                }
+            }
+        }
+
+        Ok(plugins)
+    }
+
+    fn backup(&self, marketplace: Option<&str>, name: &str) -> Result<PathBuf> {
+        let fs = RealFs;
+        let source = self.plugin_path(marketplace, name);
+        let backup_dir = self.backup_path(marketplace, name);
+
+        if !fs.exists(&source) {
+            return Err(PlmError::Cache(format!(
+                "Plugin not found: {}",
+                source.display()
+            )));
+        }
+
+        // 既存バックアップがあれば削除
+        if fs.exists(&backup_dir) {
+            fs.remove_dir_all(&backup_dir)?;
+        }
+
+        // 親ディレクトリを作成
+        if let Some(parent) = backup_dir.parent() {
+            fs.create_dir_all(parent)?;
+        }
+
+        fs.copy_dir(&source, &backup_dir)?;
+        Ok(backup_dir)
+    }
+
+    fn restore(&self, marketplace: Option<&str>, name: &str) -> Result<()> {
+        let fs = RealFs;
+        let backup_dir = self.backup_path(marketplace, name);
+        let target = self.plugin_path(marketplace, name);
+
+        if !fs.exists(&backup_dir) {
+            return Err(PlmError::Cache("Backup not found".to_string()));
+        }
+
+        // 現在のディレクトリを削除
+        if fs.exists(&target) {
+            fs.remove_dir_all(&target)?;
+        }
+
+        // バックアップからリストア
+        fs.copy_dir(&backup_dir, &target)?;
+
+        // バックアップを削除
+        fs.remove_dir_all(&backup_dir)?;
+
+        Ok(())
+    }
+
+    fn remove_backup(&self, marketplace: Option<&str>, name: &str) -> Result<()> {
+        let fs = RealFs;
+        let backup_dir = self.backup_path(marketplace, name);
+        if fs.exists(&backup_dir) {
+            fs.remove_dir_all(&backup_dir)?;
+        }
+        Ok(())
+    }
+
+    fn atomic_update(
+        &self,
+        marketplace: Option<&str>,
+        name: &str,
+        archive: &[u8],
+    ) -> Result<PathBuf> {
+        let fs = RealFs;
+        let target = self.plugin_path(marketplace, name);
+        let temp_dir = self.temp_path(marketplace, name);
+
+        // temp ディレクトリをクリーンアップ
+        if fs.exists(&temp_dir) {
+            fs.remove_dir_all(&temp_dir)?;
+        }
+
+        // 親ディレクトリを作成
+        if let Some(parent) = temp_dir.parent() {
+            fs.create_dir_all(parent)?;
+        }
+
+        // temp に展開（source_path なし）
+        extract_archive_with_source_path(&temp_dir, archive, None)?;
+
+        // 検証: plugin.json の存在確認
+        if !has_manifest(&temp_dir) {
+            fs.remove_dir_all(&temp_dir)?;
+            return Err(PlmError::InvalidManifest("plugin.json not found".into()));
+        }
+
+        // 旧キャッシュ削除 → temp をリネーム
+        if fs.exists(&target) {
+            fs.remove_dir_all(&target)?;
+        }
+
+        // リネーム（同一ファイルシステム上でのアトミック操作）
+        fs.rename(&temp_dir, &target)?;
+
+        Ok(target)
+    }
+}
+
+/// source_path の防御的検証
+///
+/// # Arguments
+///
+/// * `source_path` - optional normalized sub-path inside the archive
+fn validate_source_path(source_path: Option<&str>) -> Result<()> {
+    let sp = match source_path {
+        Some(s) => s,
+        None => return Ok(()),
+    };
+
+    if sp.contains("..") {
+        return Err(PlmError::InvalidSource(
+            "source_path is not normalized: contains '..'".into(),
+        ));
+    }
+    if sp.contains('\\') {
+        return Err(PlmError::InvalidSource(
+            "source_path is not normalized: contains backslash".into(),
+        ));
+    }
+    if sp.contains("./") || sp.starts_with('.') {
+        return Err(PlmError::InvalidSource(
+            "source_path is not normalized: contains './' or starts with '.'".into(),
+        ));
+    }
+    if Path::new(sp).is_absolute() {
+        return Err(PlmError::InvalidSource(
+            "source_path is not normalized: absolute path".into(),
+        ));
+    }
+
+    Ok(())
+}
+
+/// zipアーカイブのプレフィックス（トップディレクトリ）を取得
+///
+/// # Arguments
+///
+/// * `zip` - open zip archive to inspect
+fn get_archive_prefix(zip: &mut ZipArchive<Cursor<&[u8]>>) -> Result<String> {
+    if zip.is_empty() {
+        return Ok(String::new());
+    }
+
+    let first = zip.by_index(0)?;
+    let first_name = first.name();
+    Ok(first_name
+        .split('/')
+        .next()
+        .map(|s| format!("{}/", s))
+        .unwrap_or_default())
+}
+
+/// アーカイブを展開（source_path 指定対応）
+///
+/// # Arguments
+///
+/// * `dest` - destination directory to extract into
+/// * `archive` - zip archive bytes
+/// * `source_path` - optional normalized sub-path to extract
+fn extract_archive_with_source_path(
+    dest: &Path,
+    archive: &[u8],
+    source_path: Option<&str>,
+) -> Result<()> {
+    let cursor = Cursor::new(archive);
+    let mut zip = ZipArchive::new(cursor)?;
+
+    let prefix = get_archive_prefix(&mut zip)?;
+
+    let mut source_path_hit = false;
+    let mut entries_skipped_for_security = 0usize;
+
+    for i in 0..zip.len() {
+        let mut file = zip.by_index(i)?;
+        let file_path = file.name();
+
+        let file_path_normalized = file_path.replace('\\', "/");
+
+        let relative_path = if !prefix.is_empty() && file_path_normalized.starts_with(&prefix) {
+            &file_path_normalized[prefix.len()..]
+        } else {
+            &file_path_normalized[..]
+        };
+
+        if relative_path.is_empty() {
+            continue;
+        }
+
+        let final_path = match source_path {
+            Some(sp) => {
+                match extract_with_source_path_filter(
+                    relative_path,
+                    sp,
+                    &file,
+                    &mut source_path_hit,
+                    &mut entries_skipped_for_security,
+                ) {
+                    Some(path) => path,
+                    None => continue,
+                }
+            }
+            None => PathBuf::from(relative_path),
+        };
+
+        write_zip_entry(&mut file, &dest.join(&final_path))?;
+    }
+
+    if let Some(sp) = source_path {
+        if entries_skipped_for_security > 0 {
+            return Err(PlmError::InvalidSource(format!(
+                "{} entries in source_path were skipped for security reasons (possible zip-slip or symlink)",
+                entries_skipped_for_security
+            )));
+        }
+        if !source_path_hit {
+            return Err(PlmError::InvalidSource(format!(
+                "source_path not found in archive: {}",
+                sp
+            )));
+        }
+    }
+
+    Ok(())
+}
+
+/// source_path フィルタを適用し、展開すべきパスを返す（None = スキップ）
+///
+/// # Arguments
+///
+/// * `relative_path` - archive entry path after prefix removal
+/// * `source_path` - normalized sub-path being extracted
+/// * `file` - current zip entry (used for unix mode inspection)
+/// * `source_path_hit` - updated to `true` when a matching entry is seen
+/// * `entries_skipped` - incremented when an entry is skipped for security
+fn extract_with_source_path_filter(
+    relative_path: &str,
+    source_path: &str,
+    file: &zip::read::ZipFile,
+    source_path_hit: &mut bool,
+    entries_skipped: &mut usize,
+) -> Option<PathBuf> {
+    let relative_path_obj = Path::new(relative_path);
+    let source_path_obj = Path::new(source_path);
+
+    let stripped = relative_path_obj.strip_prefix(source_path_obj).ok()?;
+    *source_path_hit = true;
+
+    if stripped.as_os_str().is_empty() {
+        return None;
+    }
+
+    // zip-slip 対策: Normal コンポーネントのみ許容
+    let has_unsafe_component = stripped
+        .components()
+        .any(|c| !matches!(c, PathComponent::Normal(_)));
+    if has_unsafe_component {
+        *entries_skipped += 1;
+        return None;
+    }
+
+    // symlink 対策（source_path 抽出時のみ）
+    #[cfg(unix)]
+    {
+        if let Some(mode) = file.unix_mode() {
+            // S_IFLNK = 0o120000
+            if (mode & 0o170000) == 0o120000 {
+                *entries_skipped += 1;
+                return None;
+            }
+        }
+    }
+    #[cfg(not(unix))]
+    let _ = file;
+
+    Some(stripped.to_path_buf())
+}
+
+/// zipエントリをファイルシステムに書き込み
+///
+/// # Arguments
+///
+/// * `file` - current zip entry
+/// * `target_path` - destination path on the local filesystem
+fn write_zip_entry(file: &mut zip::read::ZipFile, target_path: &Path) -> Result<()> {
+    let fs = RealFs;
+    if file.is_dir() {
+        fs.create_dir_all(target_path)?;
+    } else {
+        if let Some(parent) = target_path.parent() {
+            fs.create_dir_all(parent)?;
+        }
+        let mut content = Vec::new();
+        file.read_to_end(&mut content)?;
+        fs.write(target_path, &content)?;
+    }
+    Ok(())
+}
+
+impl Default for PackageCache {
+    fn default() -> Self {
+        Self::new().expect("Failed to initialize package cache")
+    }
+}
+
+#[cfg(test)]
+#[path = "cache_test.rs"]
+mod tests;
