+++ src/plugin/cache/cleanup.rs
