+++ src/plugin/cache/cleanup_test.rs
